durationPicker.dart class moved to separate file and changed
